BE/NEAProjectBE/requirements.txt
asgiref==3.9.2
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
Faker==25.0.0
gemini-cli==0.2.5
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.62.3
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
requests==2.32.4
rich==13.9.4
rsa==4.9.1
sqlparse==0.4.4
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2023.3
urllib3==2.5.0
